+++ src/components/Lend/LoanSearch/LoanSearchSectorFilter.vue
@@ -52,8 +52,19 @@ export default {
 		}
 	},
 	methods: {
-		updateSectors(sectors) {
-			this.$emit('updated', { sectorId: sectors.map(s => +s) });
+		updateSectors({ values, changed }) {
+			this.$emit('updated', { sectorId: values.map(s => +s) });
+
+			const sector = this.displayedSectors.find(s => s.id === +changed);
+
+			if (sector) {
+				this.$kvTrackEvent?.(
+					'Lending',
+					'click-sector-filter',
+					sector.name,
+					values.includes(sector.id) ? 'selected' : 'deselected'
+				);
+			}
 		}
 	},
 	watch: {