+++ src/components/Lend/LoanSearch/LoanSearchThemeFilter.vue
@@ -50,8 +50,16 @@ export default {
 		},
 	},
 	methods: {
-		updateThemes(themes) {
-			this.$emit('updated', { theme: themes });
+		updateThemes({ values, changed }) {
+			this.$emit('updated', { theme: values });
+
+			// TODO: handle theme ID (converting to theme name) when theme ID filter used
+			this.$kvTrackEvent?.(
+				'Lending',
+				'click-theme-filter',
+				changed,
+				values.includes(changed) ? 'selected' : 'deselected'
+			);
 		}
 	},
 	watch: {