+++ src/util/flssUtils.js
@@ -27,20 +27,21 @@ export async function fetchFacets(apollo, isoCodeFilters, themeFilters, sectorFi
  * Fetches the loan data from FLSS
  *
  * @param {Object} apollo The apollo client instance
- * @param {Object} loanQueryFilters The filters for the loan query
+ * @param {Object} filterObject The filters for the loan query
  * @param {String} sortOrder Sort option for the loan query
- * @param {String} pageOffset PageNumber or Offset (FLSS uses a page number, Lend Loans uses a count offset)
+ * @param {Int} pageOffset The offset of the page
+ * @param {Int} pageLimit The limit/size of the page
  * @returns {Promise<Array<Object>>} Promise for loan data
  */
-export async function fetchLoans(apollo, loanQueryFilters, sortBy = null, pageOffset = 0) {
+export async function fetchLoans(apollo, filterObject, sortBy = null, pageOffset = 0, pageLimit = 15) {
 	try {
 		const result = await apollo.query({
 			query: flssLoanQuery,
 			variables: {
-				filterObject: loanQueryFilters,
+				filterObject,
 				sortBy,
-				pageNumber: pageOffset,
-				limit: 20
+				pageNumber: pageOffset / pageLimit,
+				pageLimit,
 			},
 			fetchPolicy: 'network-only',
 		});