+++ src/util/loanSearchUtils.js
@@ -2,6 +2,8 @@ import orderBy from 'lodash/orderBy';
 import updateLoanSearchMutation from '@/graphql/mutation/updateLoanSearchState.graphql';
 import loanFacetsQuery from '@/graphql/query/loanFacetsQuery.graphql';
 import { fetchFacets, fetchLoans } from '@/util/flssUtils';
+import { getDefaultLoanSearchState } from '@/api/localResolvers/loanSearch';
+import { isNumber } from '@/util/numberUtils';
 
 /**
  * API query types, FLSS and lend (standard)
@@ -50,23 +52,39 @@ export const sortByNameToDisplay = {
 export function getValidatedSearchState(loanSearchState, allFacets, queryType) {
 	const validSorts = queryType === FLSS_QUERY_TYPE ? allFacets.flssSorts : allFacets.standardSorts;
 
+	const defaultLoanSearchState = getDefaultLoanSearchState();
+
 	const validatedGender = allFacets.genders.includes(loanSearchState?.gender?.toUpperCase())
 		? loanSearchState.gender : null;
+
 	const validatedIsoCodes = loanSearchState?.countryIsoCode
 		?.filter(c => allFacets.countryIsoCodes.includes(c.toUpperCase())) ?? [];
+
 	const validatedSectorIds = loanSearchState?.sectorId?.filter(s => allFacets.sectorIds.includes(s)) ?? [];
+
 	const validatedSortBy = validSorts.some(s => s.name === loanSearchState.sortBy) ? loanSearchState.sortBy : null;
+
 	// TODO: change to theme IDs when theme ID filter is on prod
 	const validatedThemes = (loanSearchState?.theme?.filter(t => allFacets.themeNames.includes(t.toUpperCase())) ?? [])
 		// Handle casing differences between FLSS and lend API theme values. FLSS does case insensitive fuzzy matching.
 		.map(t => t.toUpperCase());
+
+	const validatedPageOffset = isNumber(loanSearchState?.pageOffset)
+		? loanSearchState.pageOffset
+		: defaultLoanSearchState.pageOffset;
+
+	const validatedPageLimit = isNumber(loanSearchState?.pageLimit)
+		? loanSearchState.pageLimit
+		: defaultLoanSearchState.pageLimit;
 
 	return {
 		gender: validatedGender,
 		countryIsoCode: validatedIsoCodes,
 		sectorId: validatedSectorIds,
 		sortBy: validatedSortBy,
-		theme: validatedThemes
+		theme: validatedThemes,
+		pageOffset: validatedPageOffset,
+		pageLimit: validatedPageLimit,
 	};
 }
 
@@ -349,7 +367,7 @@ export function getFlssFilters(loanSearchState) {
  * @param {Object} loanSearchState The current loan search state from Apollo
  * @returns {Object} The filter facets
  */
-export async function runFacetsQueries(apollo, loanSearchState = {}) {
+export async function runFacetsQueries(apollo, loanSearchState) {
 	const isoCodeFilters = { ...getFlssFilters(loanSearchState), countryIsoCode: undefined };
 	const themeFilters = { ...getFlssFilters(loanSearchState), theme: undefined };
 	const sectorFilters = { ...getFlssFilters(loanSearchState), sectorId: undefined };
@@ -370,14 +388,16 @@ export async function runFacetsQueries(apollo, loanSearchState = {}) {
  * @param {Object} loanSearchState The current loan search state from Apollo
  * @returns {Object} The results of the loan query
  */
-export async function runLoansQuery(apollo, loanSearchState = {}) {
+export async function runLoansQuery(apollo, loanSearchState) {
 	const flssData = await fetchLoans(
 		apollo,
 		getFlssFilters(loanSearchState),
-		loanSearchState?.sortBy ?? null
+		loanSearchState?.sortBy,
+		loanSearchState?.pageOffset,
+		loanSearchState?.pageLimit
 	);
 
-	return { loans: flssData?.values || [], totalCount: flssData?.totalCount || 0 };
+	return { loans: flssData?.values ?? [], totalCount: flssData?.totalCount ?? 0 };
 }
 
 /**
@@ -434,7 +454,7 @@ export function getSectorIdsFromQueryParam(sector, allFacets) {
 	if (!sector) return;
 
 	// Handles FLSS and legacy query params, such as "1" and "1,2"
-	if (sector.includes(',') || !Number.isNaN(Number(sector))) {
+	if (sector.includes(',') || isNumber(sector)) {
 		return sector.split(',').filter(s => s !== '').map(s => +s);
 	}
 
@@ -463,7 +483,7 @@ export function getThemeNamesQueryParam(param, facets) {
 	if (!param) return;
 
 	// Handles FLSS and legacy query params, such as "1" and "1,2"
-	if (param.includes(',') || !Number.isNaN(Number(param))) {
+	if (param.includes(',') || isNumber(param)) {
 		return param.split(',').filter(s => s !== '').reduce((prev, current) => {
 			const facet = facets.find(s => s.id === +current);
 			if (facet) {
@@ -492,14 +512,17 @@ export function getThemeNamesQueryParam(param, facets) {
  * @param {string} queryType The current query type (lend vs FLSS)
  * @param {Object} previousState The previous search state
  */
-export async function applyQueryParams(apollo, query, allFacets, queryType, previousState = {}) {
+export async function applyQueryParams(apollo, query, allFacets, queryType, previousState) {
+	// Convert query param 1-based page to pager 0-based page
+	const page = isNumber(query.page) && query.page > 0 ? query.page - 1 : 0;
+
 	const filters = {
+		...previousState, // The countryIsoCode and pageLimit are not currently in the query params
 		gender: query.gender,
 		sortBy: queryType === FLSS_QUERY_TYPE ? lendToFlssSort.get(query.sortBy) : query.sortBy,
 		sectorId: getSectorIdsFromQueryParam(query.sector, allFacets.sectorNames, allFacets.sectorFacets, true),
 		theme: getThemeNamesQueryParam(query.attribute, allFacets.themeFacets),
-		// TODO: replace previous state usage as query param support expands
-		...(previousState && { countryIsoCode: previousState.countryIsoCode })
+		pageOffset: page * previousState.pageLimit,
 	};
 
 	await updateSearchState(apollo, filters, allFacets, queryType, previousState);
@@ -542,12 +565,16 @@ export function updateQueryParams(loanSearchState, router, allFacets, queryType)
 		return prev;
 	}, []);
 
+	// Page query param is 1-based
+	const page = (loanSearchState.pageOffset / loanSearchState.pageLimit) + 1;
+
 	// Create new query params object
 	const newParams = {
 		...(loanSearchState.gender && { gender: loanSearchState.gender }),
 		...(loanSearchState.sectorId?.length && { sector: loanSearchState.sectorId.join(',') }),
 		...(loanSearchState.theme?.length && { attribute: themeIds.join(',') }),
 		...(queryParamSortBy && { sortBy: queryParamSortBy }),
+		...(page > 1 && { page: page.toString() }),
 		// TODO: add params as query param support expands
 		...utmParams,
 	};
@@ -560,6 +587,6 @@ export function updateQueryParams(loanSearchState, router, allFacets, queryType)
 
 	// Vue throws duplicate navigation exception when identical paths are pushed to the router
 	if (!doParamsMatch) {
-		router.push({ ...router.currentRoute, query: newParams, params: { noScroll: true } });
+		router.push({ ...router.currentRoute, query: newParams, params: { noScroll: true, noAnalytics: true } });
 	}
 }