+++ test/unit/specs/util/loanSearchUtils.spec.js
@@ -29,7 +29,9 @@ const mockState = {
 	countryIsoCode: ['US'],
 	sectorId: [1],
 	sortBy: 'expiringSoon',
-	theme: ['THEME 1']
+	theme: ['THEME 1'],
+	pageOffset: 10,
+	pageLimit: 5,
 };
 
 const mockAllFacets = {
@@ -150,6 +152,22 @@ describe('loanSearchUtils.js', () => {
 			const result = getValidatedSearchState(state, mockAllFacets, STANDARD_QUERY_TYPE);
 
 			expect(result).toEqual({ ...mockState, theme: [] });
+		});
+
+		it('should validate pageOffset', () => {
+			const state = { ...mockState, pageOffset: 'asd' };
+
+			const result = getValidatedSearchState(state, mockAllFacets, STANDARD_QUERY_TYPE);
+
+			expect(result).toEqual({ ...mockState, pageOffset: 0 });
+		});
+
+		it('should validate pageLimit', () => {
+			const state = { ...mockState, pageLimit: 'asd' };
+
+			const result = getValidatedSearchState(state, mockAllFacets, STANDARD_QUERY_TYPE);
+
+			expect(result).toEqual({ ...mockState, pageLimit: 15 });
 		});
 	});
 
@@ -572,8 +590,14 @@ describe('loanSearchUtils.js', () => {
 		afterEach(jest.restoreAllMocks);
 
 		it('should return loans', async () => {
-			const result = await runLoansQuery(apollo);
-			expect(spyFetchLoans).toHaveBeenCalledWith(apollo, {}, null);
+			const result = await runLoansQuery(apollo, mockState);
+			expect(spyFetchLoans).toHaveBeenCalledWith(
+				apollo,
+				getFlssFilters(mockState),
+				mockState.sortBy,
+				mockState.pageOffset,
+				mockState.pageLimit
+			);
 			expect(result).toEqual({ loans, totalCount });
 		});
 	});
@@ -774,16 +798,23 @@ describe('loanSearchUtils.js', () => {
 				variables: {
 					searchParams: {
 						gender: 'female',
-						countryIsoCode: [],
+						countryIsoCode: ['US'],
 						sectorId: [1],
 						sortBy: 'expiringSoon',
-						theme: ['THEME 1']
+						theme: ['THEME 1'],
+						pageOffset: 5,
+						pageLimit: 5,
 					}
 				}
 			};
-			const query = { ...mockState, sector: mockState.sectorId.toString(), attribute: '1' };
+			const query = {
+				...mockState,
+				sector: mockState.sectorId.toString(),
+				attribute: '1',
+				page: '2',
+			};
 
-			await applyQueryParams(apollo, query, mockAllFacets, FLSS_QUERY_TYPE);
+			await applyQueryParams(apollo, query, mockAllFacets, FLSS_QUERY_TYPE, mockState);
 
 			expect(apollo.mutate).toHaveBeenCalledWith(params);
 		});
@@ -799,6 +830,7 @@ describe('loanSearchUtils.js', () => {
 						sortBy: 'personalized',
 						sectorId: [1],
 						theme: ['THEME 1'],
+						pageOffset: 0,
 					}
 				},
 			};
@@ -825,6 +857,7 @@ describe('loanSearchUtils.js', () => {
 						sortBy: 'personalized',
 						sectorId: [],
 						theme: [],
+						pageOffset: 0,
 					}
 				},
 			};
@@ -846,6 +879,7 @@ describe('loanSearchUtils.js', () => {
 						sortBy: 'popularity',
 						sectorId: [],
 						theme: [],
+						pageOffset: 0,
 					}
 				},
 			};
@@ -856,9 +890,80 @@ describe('loanSearchUtils.js', () => {
 			expect(apollo.mutate).toHaveBeenCalledWith(params);
 		});
 
+		it('should support page', async () => {
+			const apollo = { mutate: jest.fn(() => Promise.resolve()) };
+			const params = {
+				mutation: updateLoanSearchMutation,
+				variables: {
+					searchParams: {
+						...mockState,
+						gender: null,
+						sortBy: null,
+						sectorId: [],
+						theme: [],
+						pageOffset: 15,
+					}
+				},
+			};
+			const query = { page: '4' };
+
+			await applyQueryParams(apollo, query, mockAllFacets, STANDARD_QUERY_TYPE, mockState);
+
+			expect(apollo.mutate).toHaveBeenCalledWith(params);
+		});
+
+		it('should handle negative page', async () => {
+			const apollo = { mutate: jest.fn(() => Promise.resolve()) };
+			const params = {
+				mutation: updateLoanSearchMutation,
+				variables: {
+					searchParams: {
+						...mockState,
+						gender: null,
+						sortBy: null,
+						sectorId: [],
+						theme: [],
+						pageOffset: 0,
+					}
+				},
+			};
+			const query = { page: '-1' };
+
+			await applyQueryParams(apollo, query, mockAllFacets, STANDARD_QUERY_TYPE, mockState);
+
+			expect(apollo.mutate).toHaveBeenCalledWith(params);
+		});
+
+		it('should handle non-number page', async () => {
+			const apollo = { mutate: jest.fn(() => Promise.resolve()) };
+			const params = {
+				mutation: updateLoanSearchMutation,
+				variables: {
+					searchParams: {
+						...mockState,
+						gender: null,
+						sortBy: null,
+						sectorId: [],
+						theme: [],
+						pageOffset: 0,
+					}
+				},
+			};
+			const query = { page: 'asd' };
+
+			await applyQueryParams(apollo, query, mockAllFacets, STANDARD_QUERY_TYPE, mockState);
+
+			expect(apollo.mutate).toHaveBeenCalledWith(params);
+		});
+
 		it('should not update cache when state unchanged', async () => {
 			const apollo = { mutate: jest.fn(() => Promise.resolve()) };
-			const query = { ...mockState, sector: mockState.sectorId.toString(), attribute: '1' };
+			const query = {
+				...mockState,
+				sector: mockState.sectorId.toString(),
+				attribute: '1',
+				page: '3',
+			};
 
 			await applyQueryParams(apollo, query, mockAllFacets, FLSS_QUERY_TYPE, mockState);
 
@@ -867,7 +972,7 @@ describe('loanSearchUtils.js', () => {
 	});
 
 	describe('updateQueryParams', () => {
-		it('should preserve UTM params', async () => {
+		it('should preserve UTM params', () => {
 			const state = { gender: 'female' };
 			const router = { currentRoute: { name: 'name', query: { utm_test: 'test' } }, push: jest.fn() };
 
@@ -876,20 +981,24 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { ...state, utm_test: 'test' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should push gender', async () => {
+		it('should push gender', () => {
 			const state = { gender: 'female' };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
 			updateQueryParams(state, router, mockAllFacets, FLSS_QUERY_TYPE);
 
-			expect(router.push).toHaveBeenCalledWith({ name: 'name', query: state, params: { noScroll: true } });
+			expect(router.push).toHaveBeenCalledWith({
+				name: 'name',
+				query: state,
+				params: { noScroll: true, noAnalytics: true }
+			});
 		});
 
-		it('should push sector IDs', async () => {
+		it('should push sector IDs', () => {
 			const state = { sectorId: [1, 2] };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -898,11 +1007,11 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { sector: '1,2' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should not push empty sector ID', async () => {
+		it('should not push empty sector ID', () => {
 			const state = { gender: 'female', sectorId: [] };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -911,11 +1020,11 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { gender: 'female' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should push theme IDs', async () => {
+		it('should push theme IDs', () => {
 			const state = { theme: ['THEME 1', 'THEME 2'] };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -924,11 +1033,11 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { attribute: '1,2' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should not push empty theme ID', async () => {
+		it('should not push empty theme ID', () => {
 			const state = { gender: 'female', theme: [] };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -937,11 +1046,11 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { gender: 'female' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should push mapped FLSS sort value', async () => {
+		it('should push mapped FLSS sort value', () => {
 			const state = { sortBy: 'personalized' };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -950,11 +1059,11 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { sortBy: 'popularity' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should push standard sort value', async () => {
+		it('should push standard sort value', () => {
 			const state = { sortBy: 'personalized' };
 			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
 
@@ -963,11 +1072,37 @@ describe('loanSearchUtils.js', () => {
 			expect(router.push).toHaveBeenCalledWith({
 				name: 'name',
 				query: { sortBy: 'personalized' },
-				params: { noScroll: true }
+				params: { noScroll: true, noAnalytics: true }
 			});
 		});
 
-		it('should not push identical query string', async () => {
+		it('should push page', () => {
+			const state = { pageOffset: 10, pageLimit: 2 };
+			const router = { currentRoute: { name: 'name', query: {} }, push: jest.fn() };
+
+			updateQueryParams(state, router, mockAllFacets, STANDARD_QUERY_TYPE);
+
+			expect(router.push).toHaveBeenCalledWith({
+				name: 'name',
+				query: { page: '6' },
+				params: { noScroll: true, noAnalytics: true }
+			});
+		});
+
+		it('should remove page if first page', () => {
+			const state = { pageOffset: 0, pageLimit: 2 };
+			const router = { currentRoute: { name: 'name', query: { page: '1' } }, push: jest.fn() };
+
+			updateQueryParams(state, router, mockAllFacets, STANDARD_QUERY_TYPE);
+
+			expect(router.push).toHaveBeenCalledWith({
+				name: 'name',
+				query: { },
+				params: { noScroll: true, noAnalytics: true }
+			});
+		});
+
+		it('should not push identical query string', () => {
 			const state = { gender: 'female' };
 			const router = { currentRoute: { name: 'name', query: { gender: 'female' } }, push: jest.fn() };